Tried a different approach for API accepting more that 1 entries in Database. Notsure why still not working.

+++ src/main/java/edu/wgu/assignments/MedPad/Entities/Record.java
@@ -18,13 +18,13 @@ public class Record implements Serializable {
 
     @Id
     @GeneratedValue
-    private Long id;
+    private int id;
 
     @NonNull
     private int facilityId;
 
     @NonNull
-    private Long ptId;
+    private int ptId;
 
     @NonNull
     private String chiefComplaint;
